+++ packages/zosfiles/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Zowe z/OS files SDK package will be documented in thi
 
 ## Recent Changes
 
-- BugFix: Enhanced `zowe zos-files upload file-to-uss` to support uploading a file to a USS directory using its original filename when the target is a directory (similar to UNIX cp) and added a `--make-directories` option to automatically create missing target directories during upload.[#2686](https://github.com/zowe/zowe-cli/pull/2686)
+- BugFix: Enhanced the `zowe zos-files upload file-to-uss` command to support uploading a file to a USS directory using its original filename when the target is a directory (similar to UNIX cp) and added the `--make-directories` option to automatically create missing target directories during upload.[#2686](https://github.com/zowe/zowe-cli/pull/2686)
 
 ## `8.30.2`