+++ packages/zosfiles/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe z/OS files SDK package will be documented in this file.
 
+## Recent Changes
+
+- BugFix: Added extra filesystem checks when downloading data sets and USS files. []()
+
 ## `8.32.0`
 
 - Enhancement: Added `abortDownload` predicate to `IDownloadOptions` to allow downloads in `allMembers`, `allDataSets` and `ussDir` functions to be cancelled mid-way. [#2715](https://github.com/zowe/zowe-cli/pull/2715)